refactor: load downloaded CAD models from file URLs
