+++ lib/shared/export-snippet.ts
@@ -37,7 +37,7 @@ import { convertCircuitJsonToSchematicPdf } from "./convert-circuit-json-to-sche
 import { convertToKicadLibrary } from "./convert-to-kicad-library"
 import { importFromUserLand } from "./importFromUserLand"
 import { isCircuitJsonFile } from "./is-circuit-json-file"
-import { inlineLocalCadModels } from "./inline-local-cad-models"
+import { convertModelUrlsToFileUrls } from "cli/build/convert-model-urls-to-file-urls"
 
 const writeFileAsync = promisify(fs.writeFile)
 
@@ -231,7 +231,7 @@ export const exportSnippet = async ({
     case "gltf":
       outputContent = JSON.stringify(
         await convertCircuitJsonToGltf(
-          await inlineLocalCadModels(circuitJson),
+          convertModelUrlsToFileUrls(circuitJson),
           getCircuitJsonToGltfOptions({ format: "gltf" }),
         ),
         null,
@@ -241,7 +241,7 @@ export const exportSnippet = async ({
     case "glb":
       outputContent = Buffer.from(
         (await convertCircuitJsonToGltf(
-          await inlineLocalCadModels(circuitJson),
+          convertModelUrlsToFileUrls(circuitJson),
           getCircuitJsonToGltfOptions({ format: "glb" }),
         )) as ArrayBuffer,
       )
